Accounting-Version-2/backend/requirements.txt
# Django framework
Django==5.2.4
# REST API
djangorestframework==3.15.1
# JWT authentication
djangorestframework-simplejwt==5.2.2
# Cross-Origin Resource Sharing
django-cors-headers==3.15.0
# MSSQL database backend
mssql-django==1.4
